requirements.txt
cffi==1.14.3
pycparser==2.20
PyNaCl==1.4.0
six==1.15.0
Jinja2==2.11.2